Add rounds managing system in the model
Introduces 2 main classes: ForthAndBackIterator and RoundManager.
The former handles the order in which the players can make their
move within a given round; the latter manages the correct flow of
rounds keeping track of what player is allowed to make a move at
any time. RoundManager is also responsible for detecting when a
round or the game itself ends. The class Player is only a dummy
to allow correct referencing and must be implemented.