Upload classes for Objective Cards and the PrivateObjectiveFactory
Adds the interfaces ObjectiveCards and PublicObjectiveCards, and the classes PrivateObjectiveCard, which contains the method to calculate the partial score related to it, and PrivateObjectiveFactory, which is the factory to create a given number of PrivateObjectiveCards. The class Cell has only the skeleton to allow the implementation of classes related to it, and must be implemented.