Bugfix and refactor objective cards
The static class PublicObjectiveUtils has been deleted since its use
was limited to public objectives and merged in the abstract class
PublicObjectiveCard.
The public and private objective factories now check input value and
throw an exception if it's grater than the number of unique cards
they can instantiate.
Also, an off-by-one error in the for loops in the methods numberOfShades
and numberOfColors in PublicObjectiveCard has been fixed.