Bugfix and refactor of RoundTrack, ToolCardFactory and TurnManager
Solved a bug in ToolCardFactory that allowed to produce duplicates
among the Tool Cards produced. Also RoundTrack has been modified
so that swap method now returns null when it is requested to
swap a die with a non valid position: the swap method threw
an IndexOutOfBound exception insted of returning null.
Less relevant changes has been added to TurnManager.