Refactor Cell, DraftPool, Pattern and add Enum Restriction
A copy constructor has been added to Cell to allow a deep copy of it.
The method isPresent has been implemented for DraftPool: it is
an helper method that is useful in Single-Player mode to verify
if a ToolCard is available or not. The pattern has been made
immutable in order to have a correct use of some ToolCard and some
methods have been added to achieve this task. An emptySpace method
has also been added to Pattern: it is necessary for the end Game
scoring.
The Enum Restriction represents all the possible particular restriction
that have to be respected in the placement of a Die when
using some ToolCards.