Refactor all the View messages and minor refactor of some code.
All the View messages have been updated to have an Action
(the Enum) in stead of a String representing an action as an
attribute. Also the Controller has been modified for this reason.
A method to show the difficulty's level selection in SinglePlayer
mode has been added to the View.