Refactor ToolCards 1, 2, 3, 4 and SinglePlayerController
A catch block has been added to the mentioned ToolCards to
handle the IndexOutOfBoundsException that would be thrown
if user inserts invalid Coordinates for a die placement or
removal in his pattern or for a choice from the DraftPool.
In SinglePlayerMode a control has been added to handle
the case in which the level of difficulty inserted is invalid and
to verify that Player uses the same username when reconnecting.