Implement high-level abstraction networking
Server-side, the hybrid server, the default network interface and
view-handling mechanics have been implemented and documented.
Client-side, the client and its implementation have been implemented
and documented.
The events and the part of the model that are sent over network are now
serializable, and a wrong visibility modifies on the single player
controller ha been fixed.
The new classes are all documented.