Refactor Controller class
canUseToolCard's signature has been changed: the toolCard chosen
is passed as a parameter.
The activateToolCard method has been refactored: now the selected
toolCard is assigned to the current Turn only when all the controls
are satisfied.
A new control has been added to the applyToolCard to verify that this
method is invoked only after a correct invocation to the
applyToolCard method.
The selectPattern method has been modified so that there is a complete
setup of the Game when all the players have chosen their Pattern.
Finally also the placeDie method has been modified to handle the case
in which a Player wants to place a die and the selection is forced by
the use of a toolCard.