Make some refactoring and add images for cards
The class SelectPrivateObjective of the package 'model.events' is deleted, because it was redundant, and the code related to it is refactored. Some little modifies and additions are made to the classes ClientView and Displayer in the package 'view'. The images for the color restrictions are changed and the ones for the Objective Cards are uploaded.