Move logic for Public Objective scores to the controller
The classes dealing with the calculation of the scores for the public objective cards are redesigned with a more object-oriented mentality, and moved to the controller, in order to better separate the functionalities of the elements of the MVC pattern. In the following commits the corresponding classes of the model will be adjusted, as well as the controller and the unit tests.