Refactor Objective Cards
The methods that calculate the score of the objective cards in the model are removed, the package controller is modified to accept the changes to the cards. Also tests are modified and some classes added. Also the factories for cards are modified, but some other changes must still be made.