Refactor GrozingPliersBehaviour
The View selection has been modified to allow the user to
insert the correct data when using the Grozing Pliers Tool Card.
The ClientImplementation and all the necesssary View classes
have been modified as a consequence to handle this new View selection
and to display it in a correct way.