Add classes to load and instantiate PublicObjectiveCards from files
The classes SaxPublicObjectiveBuilder and XmlPublicObjectiveLoader are added, and consequently many classes of the package 'controller' are modified. Tests are corrected, adding a directory that contains some xmls to test the loading of PublicObjectiveCards. The xml for those cards are corrected, as well as the xsd. Other refactoring is made to tests, in order to solve sonar issues that had been neglected.