Move the PublicObjectiveFactory to the tests
The old PublicObjectiveFactory, now substituted by the XmlPublicObjectiveLoader, is moved to the test package, in order not to ruin some tests, in which having the old factory or the new one doesn't make difference. The tests in which there is a difference are changed. Some constants in the classes of the controller are declared at the beginning of the class instead of using them in the middle of the code, to make code cleaner.