Refactor tool card usage system
Common code was grouped to remove possible duplications.
Also, ToolCardBehaviours are no longer classified by tool card name, but
rather by their effects and usage mechanics.
Tests and documentation were added and modified to fit the new
design.
The event message MoveDie was removed since it represented a subset of
information of MoveDice and the event message IncrementDieValue is now
a subclass of SelectDie to allow to abstract code.