Fix bugs for single player mode
Tool cards used to handle poorly the sacrifice die used to activate
them, which led to incorrect display and bad behaviour.
In single player mode, the game was finalized too early, preventing the
score board to be sent to clients.
Reconnection for single player mode has been removed because it had too
many bugs and is not required.
Involved tests have been adjusted.