Fix display bug
In GUI mode, selecting a die made disabled dice coloured again.
In both CLI and GUI mode is now possible to see if a tool card has
already been used. Also, the player is informed if it's his first or
second turn this round.
A buf in ViewDataOrganizer prevented a player from resulting connected
after the first disconnection.